refactor: split to regex file
