+++ src/smt/CMakeLists.txt
@@ -76,6 +76,7 @@ z3_add_component(smt
     theory_str_noodler/nielsen_decision_procedure.cpp
     theory_str_noodler/formula.cpp
     theory_str_noodler/util.cc
+    theory_str_noodler/regex.cpp
     theory_str_noodler/aut_assignment.cpp
     theory_str_mc.cpp
     theory_str_regex.cpp
