+++ src/smt/theory_str_noodler/theory_str_noodler.h
@@ -34,6 +34,7 @@ Eternal glory to Yu-Fang.
 #include "decision_procedure.h"
 #include "expr_solver.h"
 #include "util.h"
+#include "regex.h"
 #include "var_union_find.h"
 #include "nielsen_decision_procedure.h"
