+++ src/smt/theory_str_noodler/util.h
@@ -34,20 +34,6 @@ namespace smt::noodler::util {
     using expr_pair_flag = std::tuple<expr_ref, expr_ref, bool>;
 
     /**
-     * @brief Info gathered about a regex. 
-     * - min_length: length of shortest words in the regex. In fact it expresses that in the regex there is no 
-     *      word shorter than min_length. It does not mean that regex contains a word of length exactly min_length. 
-     *      If empty == l_true or l_undef, this values is not valid. 
-     * - universal: is regex universal?
-     * - empty: is regex empty?
-     */
-    struct RegexInfo {
-        unsigned min_length;
-        lbool universal;
-        lbool empty;
-    };
-
-    /**
      * Throws error and select which class to throw based on debug (if we are
      * debugging, we do not want z3 to catch our error, if we are not debugging
      * we want z3 to catch it and return unknown).
@@ -95,15 +81,6 @@ namespace smt::noodler::util {
     void get_variable_names(expr* ex, const seq_util& m_util_s, const ast_manager& m, std::unordered_set<std::string>& res);
 
     /**
-     * Extract symbols from a given expression @p ex. Append to the output parameter @p alphabet.
-     * @param[in] ex Expression to be checked for symbols.
-     * @param[in] m_util_s Seq util for AST.
-     * @param[in] m AST manager.
-     * @param[out] alphabet A set of symbols with where found symbols are appended to.
-     */
-    void extract_symbols(expr * ex, const seq_util& m_util_s, const ast_manager& m, std::set<uint32_t>& alphabet);
-
-    /**
      * Get dummy symbols.
      *
      * @param[in] new_symb_num Number of added symbols.
@@ -111,26 +88,6 @@ namespace smt::noodler::util {
      * @return Set of dummy symbols.
      */
     std::set<uint32_t> get_dummy_symbols(size_t new_symb_num, std::set<uint32_t>& symbols_to_append_to);
-
-    /**
-     * Convert expression @p expr to NFA using hexadecimal values as symbols.
-     * @param[in] expression Expression to be converted to regex.
-     * @param[in] m_util_s Seq util for AST.
-     * @param[in] m AST manager.
-     * @param[in] alphabet Alphabet to be used in re.allchar (SMT2: '.') expressions.
-     * @param[in] determinize Determinize intermediate automata
-     * @param[in] make_complement Whether to make complement of the passed @p expr instead.
-     * @return The resulting regex.
-     */
-    [[nodiscard]] Mata::Nfa::Nfa conv_to_nfa(const app *expression, const seq_util& m_util_s, const ast_manager& m,
-                                             const std::set<uint32_t>& alphabet, bool determinize = false, bool make_complement = false);
-
-    /**
-     * Create NFA accepting a word in Z3 zstring representation.
-     * @param word Word to accept.
-     * @return NFA.
-     */
-    Mata::Nfa::Nfa create_word_nfa(const zstring& word);
 
     /**
      * Collect basic terms (vars, literals) from a concatenation @p ex. Append the basic terms to the output parameter
@@ -240,16 +197,6 @@ namespace smt::noodler::util {
      */
     expr_ref len_to_expr(const LenNode &node, const std::map<BasicTerm, expr_ref>& variable_map, ast_manager &m, seq_util& m_util_s, arith_util& m_util_a);
 
-    /**
-     * @brief Get basic information about the regular expression in the form of RegexInfo (see the description above). 
-     * RegexInfo gathers information about emptyness; universality; length of shortest words
-     * 
-     * @param expression Regex to be checked
-     * @param m_util_s tring ast util
-     * @param m ast manager
-     * @return RegexInfo 
-     */
-    RegexInfo get_regex_info(const app *expression, const seq_util& m_util_s, const ast_manager& m);
 }
 
 #endif
